app/requirements.txt
djangorestframework==3.10.2
simplejson==3.16.0
django-debug-toolbar==2.0
python-dateutil==2.8.*
reportlab==3.5.*
django-ipware==2.1.*
boto3
requests==2.23.0
Django==2.2.9

djangorestframework-simplejwt==4.4.0
django-cors-headers==3.2.1